abstractvar.py
