+++ pyteal/ast/__init__.py
@@ -165,7 +165,8 @@ from pyteal.ast.scratch import (
     ScratchStackStore,
     ScratchStore,
 )
-from pyteal.ast.scratchvar import DynamicScratchVar, ScratchVar, AbstractVar
+from pyteal.ast.scratchvar import DynamicScratchVar, ScratchVar
+from pyteal.ast.abstractvar import AbstractVar
 from pyteal.ast.maybe import MaybeValue
 from pyteal.ast.multi import MultiValue
 from pyteal.ast.opup import OpUp, OpUpMode, OpUpFeeSource
