+++ pyteal/ast/scratchvar.py
@@ -1,27 +1,9 @@
-from abc import ABC, abstractmethod
-
 from pyteal.errors import TealInputError
 from pyteal.types import TealType, require_type, types_match
 
+from pyteal.ast.abstractvar import AbstractVar
 from pyteal.ast.expr import Expr
 from pyteal.ast.scratch import ScratchSlot, ScratchLoad, ScratchStore
-
-
-class AbstractVar(ABC):
-    @abstractmethod
-    def store(self, value: Expr) -> Expr:
-        pass
-
-    @abstractmethod
-    def load(self) -> Expr:
-        pass
-
-    @abstractmethod
-    def storage_type(self) -> TealType:
-        pass
-
-
-AbstractVar.__module__ = "pyteal"
 
 
 class ScratchVar(AbstractVar):
